Add deploy command, remove build_serverless_bundle command
